Update gradle files
This converts a few of the build files to use the Kotlin DSL instead of
the Groovy one. The Kotlin one is newer and I'd like to use it since the
rest of the project is Kotlin.

The only one that isn't yet is the lwjgl3 project, because when I
attempted to update to Kotlin it stopped working and I couldn't find the
Kotlin equivalents for some of what's in there yet.