Box2D functionality in OverWorldScreen
This is a whole ton of functionality all in one. I'm adding in my new
systems and removing using the old ones that aren't needed anymore.

It also adds the world, and adds the character to the world (with some
gross code).

And finally it uses the tiled map object layer to get the objects that
are the borders and add them to the map. For now no entities for these,
as I just need to make sure that the world knows about them.