Switch GrowthStage to class
Instead of having this be an enum, I want to be able to support any
number of growth stages for a crop. Not all crops are going to have just
2 stages, some might have even more.

This makes all of this much easier, and while possibly being a bit more
confusing to configure, I think will give more flexibility in the long
run.

Eventually I want these to be able to be configured via JSON config
files.