Use extension functions for components
Constantly having to use the mapper was making for some not amazing
feeling code to write. Whereas this feels much better to just go to the
entity and get the component out of it. Still just buckets of data, but
at least getting the buckets is easier.

Right now almost all of them will fail if they are ever null, but I hope
I can leave that in place.